+++ src/extension.ts
@@ -2,7 +2,7 @@
 import * as which from 'which';
 import * as vscode from 'vscode';
 
-import FortranLintingProvider from './features/linter-provider';
+import { FortranLintingProvider } from './features/linter-provider';
 import FortranHoverProvider from './features/hover-provider';
 import { FortranCompletionProvider } from './features/completion-provider';
 import { FortranDocumentSymbolProvider } from './features/document-symbol-provider';
@@ -11,58 +11,56 @@ import { LoggingService } from './services/logging-service';
 import * as pkg from '../package.json';
 import { LANG_SERVER_TOOL_ID } from './lib/tools';
 import { FortranFormattingProvider } from './features/formatting-provider';
+import { FortranLanguageServer } from './fortls-interface';
 import { EXTENSION_ID, FortranDocumentSelector, promptForMissingTool } from './lib/tools';
 
 // Make it global to catch errors when activation fails
 const loggingService = new LoggingService();
 
 export function activate(context: vscode.ExtensionContext) {
-  const extensionConfig = vscode.workspace.getConfiguration(EXTENSION_ID);
+  const config = vscode.workspace.getConfiguration(EXTENSION_ID);
+  const linterType = config.get<string>('linter.compiler');
+  const formatterType = config.get<string>('formatting.formatter');
+  const autocompleteType = config.get<string>('provide.autocomplete');
+  const hoverType = config.get<string>('provide.hover');
+  const symbolsType = config.get<string>('provide.symbols');
+  detectDeprecatedOptions();
 
   loggingService.logInfo(`Extension Name: ${pkg.displayName}`);
   loggingService.logInfo(`Extension Version: ${pkg.version}`);
+  loggingService.logInfo(`Linter set to: ${linterType}`);
+  loggingService.logInfo(`Formatter set to: ${formatterType}`);
+  loggingService.logInfo(`Autocomplete set to: ${autocompleteType}`);
+  loggingService.logInfo(`Hover set to: ${hoverType}`);
+  loggingService.logInfo(`Symbols set to: ${symbolsType}`);
 
-  if (extensionConfig.get('linterEnabled', true)) {
+  if (linterType !== 'Disabled') {
     const linter = new FortranLintingProvider(loggingService);
     linter.activate(context.subscriptions);
     vscode.languages.registerCodeActionsProvider(FortranDocumentSelector(), linter);
-    loggingService.logInfo('Linter is enabled');
-  } else {
-    loggingService.logInfo('Linter is not enabled');
   }
 
-  if (extensionConfig.get('formatter') !== 'Disabled') {
+  if (formatterType !== 'Disabled') {
     const disposable: vscode.Disposable = vscode.languages.registerDocumentFormattingEditProvider(
       FortranDocumentSelector(),
       new FortranFormattingProvider(loggingService)
     );
     context.subscriptions.push(disposable);
-    loggingService.logInfo('Formatting is enabled');
-  } else {
-    loggingService.logInfo('Formatting is disabled');
   }
 
-  if (extensionConfig.get('provideCompletion', true)) {
+  if (autocompleteType === 'Built-in') {
     const completionProvider = new FortranCompletionProvider(loggingService);
     vscode.languages.registerCompletionItemProvider(FortranDocumentSelector(), completionProvider);
-  } else {
-    loggingService.logInfo('Completion Provider is not enabled');
   }
 
-  if (extensionConfig.get('provideHover', true)) {
+  if (hoverType === 'Built-in') {
     const hoverProvider = new FortranHoverProvider(loggingService);
     vscode.languages.registerHoverProvider(FortranDocumentSelector(), hoverProvider);
-    loggingService.logInfo('Hover Provider is enabled');
-  } else {
-    loggingService.logInfo('Hover Provider is not enabled');
   }
 
-  if (extensionConfig.get('provideSymbols', true)) {
+  if (symbolsType === 'Built-in') {
     const symbolProvider = new FortranDocumentSymbolProvider();
     vscode.languages.registerDocumentSymbolProvider(FortranDocumentSelector(), symbolProvider);
-    loggingService.logInfo('Symbol Provider is enabled');
-  } else {
-    loggingService.logInfo('Symbol Provider is not enabled');
   }
 
   // Check if the language server is installed and if not prompt to install it
@@ -72,5 +70,33 @@ export function activate(context: vscode.ExtensionContext) {
               For a full list of features the language server adds see:
               https://github.com/hansec/fortran-language-server`;
     promptForMissingTool(LANG_SERVER_TOOL_ID, msg, 'Python', loggingService);
+  }
+
+  // Spawn the fortran-language-server
+  const fortls = new FortranLanguageServer(loggingService);
+  fortls.activate(context.subscriptions);
+}
+
+function detectDeprecatedOptions() {
+  const config = vscode.workspace.getConfiguration(EXTENSION_ID);
+  const oldArgs: string[] = [];
+  if (config.get('includePaths')) oldArgs.push('fortran.includePaths');
+  if (config.get('gfortranExecutable')) oldArgs.push('fortran.gfortranExecutable');
+  if (config.get('linterEnabled')) oldArgs.push('fortran.linterEnabled');
+  if (config.get('linterExtraArgs')) oldArgs.push('fortran.linterExtraArgs');
+  if (config.get('linterModOutput')) oldArgs.push('fortran.linterModOutput');
+  if (config.get('symbols')) oldArgs.push('fortran.symbols');
+  if (config.get('provideSymbols')) oldArgs.push('fortran.provideSymbols');
+  if (config.get('provideHover')) oldArgs.push('fortran.provideHover');
+  if (config.get('provideCompletion')) oldArgs.push('fortran.provideCompletion');
+
+  if (oldArgs.length !== 0) {
+    const err = vscode.window.showErrorMessage(
+      `Deprecated settings have been detected in your settings.
+       Please update your settings to make use of the new names. The old names will not work.`,
+      'Open Settings'
+    );
+    err.then(() => vscode.commands.executeCommand('workbench.action.openGlobalSettings'));
+    loggingService.logError(`The following deprecated options have been detected:\n${oldArgs}`);
   }
 }