+++ src/fortls-interface.ts
@@ -1,0 +1,173 @@
+// Modified version of Chris Hansen's Fortran Intellisense
+
+'use strict';
+
+import { spawnSync } from 'child_process';
+import { commands, window, workspace, TextDocument, Disposable } from 'vscode';
+import { LanguageClient, LanguageClientOptions, ServerOptions } from 'vscode-languageclient/node';
+import { EXTENSION_ID, FortranDocumentSelector } from './lib/tools';
+import { LoggingService } from './services/logging-service';
+
+export class FortranLanguageServer {
+  constructor(private logger: LoggingService) {
+    this.logger.logInfo('Fortran Language Server');
+  }
+
+  private clients: Map<string, LanguageClient> = new Map();
+  private _fortlsVersion: string | undefined;
+
+  public activate(disposables: Disposable[]) {
+    workspace.onDidOpenTextDocument(this.didOpenTextDocument, this, disposables);
+    workspace.textDocuments.forEach(this.didOpenTextDocument, this);
+    return;
+  }
+
+  public deactivate() {
+    const promises: Thenable<void>[] = [];
+    for (const client of this.clients.values()) {
+      promises.push(client.stop());
+    }
+    return undefined;
+  }
+
+  /**
+   * Launch the fortran-language-server. This launches the server for all
+   * glob matches in the workspace and caches it.
+   *
+   * for references on the method see:
+   * https://code.visualstudio.com/api/language-extensions/language-server-extension-guide#implementing-a-language-server
+   *
+   * @param document document to lint
+   * @returns
+   */
+  private didOpenTextDocument(document: TextDocument): void {
+    // We are only interested in Fortran files
+    if (!document.languageId.toLowerCase().startsWith('fortran')) {
+      return;
+    }
+    const uri = document.uri;
+    const folder = workspace.getWorkspaceFolder(uri);
+    // Files outside a folder can't be handled. This might depend on the language.
+    // Single file languages like JSON might handle files outside the workspace folders.
+    if (!folder) {
+      return;
+    }
+    this.logger.logInfo('Fortran Language Server');
+
+    // If we have nested workspace folders we only start a server on the outer most workspace folder.
+    // folder = this.getOuterMostWorkspaceFolder(folder);
+    if (this.clients.has(folder.uri.toString())) {
+      return;
+    }
+
+    // Get path for the language server
+    const conf = workspace.getConfiguration(EXTENSION_ID);
+    const executablePath = conf.get<string>('path') || 'fortls';
+    const maxLineLength = conf.get<number>('maxLineLength') || -1;
+    const maxCommentLineLength = conf.get<number>('maxCommentLineLength') || -1;
+
+    // Setup server arguments
+    const args: string[] = ['--enable_code_actions'];
+    // todo: fortran.useFortranLanguageServer: true|false
+    if (conf.get<string>('provide.autocomplete', 'fortls') === 'Disabled') {
+      args.push('--autocomplete_no_prefix');
+    }
+    if (conf.get<string>('preferredCase', 'lowercase') === 'lowercase') {
+      args.push('--lowercase_intrinsics');
+    }
+    if (conf.get<string>('provide.symbols', 'fortls') === 'Disabled') {
+      args.push('--symbol_skip_mem');
+    }
+
+    // Enable all hover functionality. Does not make much sense to have one
+    // but not the other two
+    if (
+      conf.get<string>('provide.hover', 'fortls') === 'fortls' ||
+      conf.get<string>('provide.hover', 'fortls') === 'Both'
+    ) {
+      args.push('--variable_hover', '--hover_signature', '--use_signature_help');
+    }
+
+    // FORTLS specific args with no overlap with the main extension
+    if (conf.get<boolean>('fortls.preserveKeywordOrder')) {
+      args.push('--preserve_keyword_order');
+    }
+    if (conf.get<boolean>('fortls.disableDiagnostics')) {
+      args.push('--disable_diagnostics');
+    }
+    if (conf.get<boolean>('fortls.incrementalSync')) {
+      args.push('--incrmental_sync');
+    }
+    if (conf.get<boolean>('fortls.notifyInit')) {
+      args.push('--notify_init');
+    }
+
+    // TODO: this should also be passed in the linter.extraArgs
+    if (maxLineLength > 0) {
+      args.push(`--max_line_length=${maxLineLength}`);
+    }
+    if (maxCommentLineLength > 0) {
+      args.push(`--max_comment_line_length=${maxCommentLineLength}`);
+    }
+
+    // Detect language server version and verify selected options
+    this._fortlsVersion = this.getLSVersion(executablePath, args);
+    if (this._fortlsVersion) {
+      // Check if there is a newer version
+      //  this.checkVersion(this._fortlsVersion);
+
+      const serverOptions: ServerOptions = {
+        command: executablePath,
+        args: args,
+      };
+
+      // Options to control the language client
+      const clientOptions: LanguageClientOptions = {
+        // Register the server for all Fortran documents in workspace
+        documentSelector: FortranDocumentSelector(folder),
+        workspaceFolder: folder,
+      };
+
+      // Create the language client, start the client and add it to the registry
+      const client = new LanguageClient(
+        'fortls',
+        'Fortran Language Server',
+        serverOptions,
+        clientOptions
+      );
+      client.start();
+      this.clients.set(folder.uri.toString(), client);
+    }
+  }
+
+  /**
+   * Check if `fortls` is present and the arguments being passed are correct
+   * The presence check has already been done in the extension activate call
+   *
+   * Code taken from Fortran Intellisense
+   *
+   * @param executablePath path for `fortls` if not in path  else `fortls`
+   * @param args `fortls` input arguments see documentation
+   * @returns truthy value on success
+   */
+  private getLSVersion(executablePath: string, args: string[]) {
+    const results = spawnSync(executablePath, args.concat(['--version']));
+    if (results.error) {
+      const selected = window.showErrorMessage(
+        'Error spawning fortls: Please check that fortran-language-server is installed and in your path.',
+        'Open settings'
+      );
+      selected.then(() => commands.executeCommand('workbench.action.openGlobalSettings'));
+      return null;
+    }
+    if (results.status !== 0) {
+      const selected = window.showErrorMessage(
+        'Error launching fortls: Please check that all selected options are supported by your language server version.',
+        'Open settings'
+      );
+      selected.then(() => commands.executeCommand('workbench.action.openGlobalSettings'));
+      return null;
+    }
+    return results.stdout.toString().trim();
+  }
+}